feat(cli): add output format for bootstrap
