+++ internal/cliutils/setupclientcontext.go
@@ -63,7 +63,7 @@ func RequireBootstrapped(ctx context.Context, cfg *rest.Config, rawCfg *api.Conf
 			ExitWithError()
 		}
 		client := bootstrap.NewBootstrapClient(cfg)
-		if err := client.Bootstrap(ctx, bootstrap.DefaultOptions(), ""); err != nil {
+		if _, err := client.Bootstrap(ctx, bootstrap.DefaultOptions()); err != nil {
 			fmt.Fprintf(os.Stderr, "\nAn error occurred during bootstrap:\n%v\n", err)
 			ExitWithError()
 		} else {
