+++ pkg/bootstrap/bootstrap.go
@@ -5,10 +5,7 @@ import (
 	"encoding/json"
 	"fmt"
 	"net/http"
-	"os"
 	"time"
-
-	js "k8s.io/apimachinery/pkg/runtime/serializer/json"
 
 	"github.com/fatih/color"
 	"github.com/glasskube/glasskube/api/v1alpha1"
@@ -27,7 +24,6 @@ import (
 	"k8s.io/apimachinery/pkg/api/meta"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
-	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/client-go/discovery"
 	"k8s.io/client-go/dynamic"
@@ -74,15 +70,15 @@ func (c *BootstrapClient) initRestMapper() error {
 	}
 }
 
-func (c *BootstrapClient) Bootstrap(ctx context.Context, options BootstrapOptions, output string) error {
+func (c *BootstrapClient) Bootstrap(ctx context.Context, options BootstrapOptions) ([]unstructured.Unstructured, error) {
 	telemetry.BootstrapAttempt()
 
 	if err := c.initRestMapper(); err != nil {
-		return err
+		return nil, err
 	}
 
 	if client, err := dynamic.NewForConfig(c.clientConfig); err != nil {
-		return err
+		return nil, err
 	} else {
 		c.client = client
 	}
@@ -96,10 +92,10 @@ func (c *BootstrapClient) Bootstrap(ctx context.Context, options BootstrapOption
 			if releaseInfo, err := releaseinfo.FetchLatestRelease(); err != nil {
 				if httperror.Is(err, http.StatusServiceUnavailable) || httperror.IsTimeoutError(err) {
 					telemetry.BootstrapFailure(time.Since(start))
-					return fmt.Errorf("network connectivity error, check your network: %w", err)
+					return nil, fmt.Errorf("network connectivity error, check your network: %w", err)
 				}
 				telemetry.BootstrapFailure(time.Since(start))
-				return fmt.Errorf("could not determine latest version: %w", err)
+				return nil, fmt.Errorf("could not determine latest version: %w", err)
 			} else {
 				version = releaseInfo.Version
 			}
@@ -113,7 +109,7 @@ func (c *BootstrapClient) Bootstrap(ctx context.Context, options BootstrapOption
 	if err != nil {
 		statusMessage("Couldn't fetch Glasskube manifests", false)
 		telemetry.BootstrapFailure(time.Since(start))
-		return err
+		return nil, err
 	}
 
 	statusMessage("Validating existing installation", true)
@@ -122,7 +118,7 @@ func (c *BootstrapClient) Bootstrap(ctx context.Context, options BootstrapOption
 		telemetry.BootstrapFailure(time.Since(start))
 		statusMessage(fmt.Sprintf("Couldn't prepare manifests: %v", err), false)
 		if !options.Force {
-			return err
+			return nil, err
 		} else {
 			statusMessage("Attempting to force bootstrap anyways (Force option is enabled)", true)
 		}
@@ -137,52 +133,14 @@ func (c *BootstrapClient) Bootstrap(ctx context.Context, options BootstrapOption
 	if err = c.applyManifests(ctx, manifests); err != nil {
 		telemetry.BootstrapFailure(time.Since(start))
 		statusMessage(fmt.Sprintf("Couldn't apply manifests: %v", err), false)
-		return err
+		return nil, err
 	}
 
 	elapsed := time.Since(start)
 	c.handleTelemetry(options.DisableTelemetry, elapsed)
 
 	statusMessage(fmt.Sprintf("Glasskube successfully installed! (took %v)", elapsed.Round(time.Second)), true)
-	if output != "" {
-		if err := convertAndPrintManifests(manifests, output); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func convertAndPrintManifests(
-	objs []unstructured.Unstructured,
-	output string,
-) error {
-	scheme := runtime.NewScheme()
-	var opt bool
-	var div string
-	switch output {
-	case "json":
-		opt = false
-		div = ","
-	case "yaml":
-		opt = true
-		div = "---"
-	}
-	serializer := js.NewSerializerWithOptions(
-		js.DefaultMetaFactory, scheme, scheme,
-		js.SerializerOptions{Yaml: opt, Pretty: true, Strict: true},
-	)
-
-	for _, obj := range objs {
-		runtimeObj := &unstructured.Unstructured{}
-		obj.DeepCopyInto(runtimeObj)
-
-		if err := serializer.Encode(runtimeObj, os.Stdout); err != nil {
-			return fmt.Errorf("failed to serialize object %v: %v", obj.GetName(), err)
-		}
-		fmt.Println(div)
-	}
-
-	return nil
+	return manifests, nil
 }
 
 func (c *BootstrapClient) preprocessManifests(
